Reset cursor in zsh config instead of vim
When exiting neovim, the cursor tends to get stuck in block mode, which
is quite annoying. My fix for now was to set an autocmd in the nvim
config, but when running neovim without its config, the cursor still was
a block on exit (obviously). So what I did is add a hook to the zsh
config to re-set the cursor whenever a command exits. That way it works
even if you get out of a root shell (when getting back to the user
shell, which has the zsh config, the cursor will still be reset).