Only search with depth 1 when looking for a repo
I changed the search for those 2 scripts so that they only show the
directories in the root. That way anything can be a repo (even when not
initiated with git) and we don't end up with all the sub-repos
(submodules or others). This has a tradeoff: we can't find repos that
are in subdirectories, but I think that's alright.