Possibly fix font
For quite some time now, there has been a bug when installing the .f's
on a new system where some symbols do not show right in the wibar. I
hope it's because of the installed font package, and I've updated it to
the only nerd font that is currently installed on my system:
nerd-fonts-sf-mono-ligatures.
I will have to check it's the right one later.