aWM: Fix formatting for mpd and server load in bar
Before this commit, the formatting wasn't consistent, with the server icon
being just next to the load and the mpd status being two spaces away from the
icon. This commit changes this to both having one space between the visual icon
and the text (the icon character being 2 characters wide, we have 2 spaces
between the icon character and the text).